Add Accordion transition
1. Why is this change necessary?
Because the main project has this functionality

2. How does it address the issue?
This adds `react-bootstrap` dependency to the project and use the
`Accordion` component that handles the smooth transition

3. What side effects does this change have?
None
