+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "react": "^16.12.0",
+    "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.12.0",
     "react-scripts": "3.2.0"
   },
